Move daemon command to DaemonCommand class
